server/requirements.txt
six==1.15.0
bs4==0.0.1
tqdm==4.54.0
redis==3.5.3
image==1.5.33
boto3==1.16.9
bcrypt
daphne==3.0.0
django==3.0.5
moviepy==1.0.0
sqlparse==0.2.4
pytest==6.1.2
ipcalc==1.99.0
demjson==2.2.4
requests==2.22.0
botocore==1.19.13
virtualenv==20.1.0
mongoengine==0.20.0
coloredlogs==14.0
django-redis==4.12.1
python-dotenv==0.15.0
django-crontab==0.7.1
awesome-slugify==1.6.5
instagram-scraper
python-telegram-bot==13.0
django-rest-framework-mongoengine==3.4.1